+++ command/agent/config.go
@@ -21,14 +21,16 @@ const DefaultBindPort int = 7946
 // DefaultConfig contains the defaults for configurations.
 func DefaultConfig() *Config {
 	return &Config{
-		Tags:          make(map[string]string),
-		BindAddr:      "0.0.0.0",
-		AdvertiseAddr: "",
-		LogLevel:      "INFO",
-		RPCAddr:       "127.0.0.1:7373",
-		Protocol:      serf.ProtocolVersionMax,
-		ReplayOnJoin:  false,
-		Profile:       "lan",
+		Tags:           make(map[string]string),
+		BindAddr:       "0.0.0.0",
+		AdvertiseAddr:  "",
+		LogLevel:       "INFO",
+		RPCAddr:        "127.0.0.1:7373",
+		Protocol:       serf.ProtocolVersionMax,
+		ReplayOnJoin:   false,
+		Profile:        "lan",
+		RetryInterval:  30 * time.Second,
+		SyslogFacility: "LOCAL0",
 	}
 }
 
@@ -50,6 +52,11 @@ type Config struct {
 	// the 'role' key is special, and is used for backwards compatibility.
 	Tags map[string]string `mapstructure:"tags"`
 
+	// TagsFile is the path to a file where Serf can store its tags. Tag
+	// persistence is desirable since tags may be set or deleted while the
+	// agent is running. Tags can be reloaded from this file on later starts.
+	TagsFile string `mapstructure:"tags_file"`
+
 	// BindAddr is the address that the Serf agent's communication ports
 	// will bind to. Serf will use this address to bind to for both TCP
 	// and UDP connections. If no port is present in the address, the default
@@ -67,6 +74,10 @@ type Config struct {
 	// "head -c16 /dev/urandom | base64". If this is not specified, the
 	// traffic will not be encrypted.
 	EncryptKey string `mapstructure:"encrypt_key"`
+
+	// KeyringFile is the path to a file containing a serialized keyring.
+	// The keyring is used to facilitate encryption.
+	KeyringFile string `mapstructure:"keyring_file"`
 
 	// LogLevel is the level of the logs to output.
 	// This can be updated during a reload.
@@ -147,6 +158,41 @@ type Config struct {
 	// by having the minority node shutdown. If you want to disable this behavior,
 	// then this flag can be set to true.
 	DisableNameResolution bool `mapstructure:"disable_name_resolution"`
+
+	// EnableSyslog is used to also tee all the logs over to syslog. Only supported
+	// on linux and OSX. Other platforms will generate an error.
+	EnableSyslog bool `mapstructure:"enable_syslog"`
+
+	// SyslogFacility is used to control which syslog facility messages are
+	// sent to. Defaults to LOCAL0.
+	SyslogFacility string `mapstructure:"syslog_facility"`
+
+	// RetryJoin is a list of addresses to attempt to join when the
+	// agent starts. Serf will continue to retry the join until it
+	// succeeds or RetryMaxAttempts is reached.
+	RetryJoin []string `mapstructure:"retry_join"`
+
+	// RetryMaxAttempts is used to limit the maximum attempts made
+	// by RetryJoin to reach other nodes. If this is 0, then no limit
+	// is imposed, and Serf will continue to try forever. Defaults to 0.
+	RetryMaxAttempts int `mapstructure:"retry_max_attempts"`
+
+	// RetryIntervalRaw is the string retry interval. This interval
+	// controls how often we retry the join for RetryJoin. This defaults
+	// to 30 seconds.
+	RetryIntervalRaw string        `mapstructure:"retry_interval"`
+	RetryInterval    time.Duration `mapstructure:"-"`
+
+	// RejoinAfterLeave controls our interaction with the snapshot file.
+	// When set to false (default), a leave causes a Serf to not rejoin
+	// the cluster until an explicit join is received. If this is set to
+	// true, we ignore the leave, and rejoin the cluster on start. This
+	// only has an affect if the snapshot file is enabled.
+	RejoinAfterLeave bool `mapstructure:"rejoin_after_leave"`
+
+	// StatsiteAddr is the address of a statsite instance. If provided,
+	// metrics will be streamed to that instance.
+	StatsiteAddr string `mapstructure:"statsite_addr"`
 }
 
 // BindAddrParts returns the parts of the BindAddr that should be
@@ -211,8 +257,9 @@ func DecodeConfig(r io.Reader) (*Config, error) {
 	var md mapstructure.Metadata
 	var result Config
 	msdec, err := mapstructure.NewDecoder(&mapstructure.DecoderConfig{
-		Metadata: &md,
-		Result:   &result,
+		Metadata:    &md,
+		Result:      &result,
+		ErrorUnused: true,
 	})
 	if err != nil {
 		return nil, err
@@ -245,6 +292,14 @@ func DecodeConfig(r io.Reader) (*Config, error) {
 			return nil, err
 		}
 		result.TombstoneTimeout = dur
+	}
+
+	if result.RetryIntervalRaw != "" {
+		dur, err := time.ParseDuration(result.RetryIntervalRaw)
+		if err != nil {
+			return nil, err
+		}
+		result.RetryInterval = dur
 	}
 
 	return &result, nil
@@ -335,6 +390,30 @@ func MergeConfig(a, b *Config) *Config {
 	if b.DisableNameResolution {
 		result.DisableNameResolution = true
 	}
+	if b.TagsFile != "" {
+		result.TagsFile = b.TagsFile
+	}
+	if b.KeyringFile != "" {
+		result.KeyringFile = b.KeyringFile
+	}
+	if b.EnableSyslog {
+		result.EnableSyslog = true
+	}
+	if b.RetryMaxAttempts != 0 {
+		result.RetryMaxAttempts = b.RetryMaxAttempts
+	}
+	if b.RetryInterval != 0 {
+		result.RetryInterval = b.RetryInterval
+	}
+	if b.RejoinAfterLeave {
+		result.RejoinAfterLeave = true
+	}
+	if b.SyslogFacility != "" {
+		result.SyslogFacility = b.SyslogFacility
+	}
+	if b.StatsiteAddr != "" {
+		result.StatsiteAddr = b.StatsiteAddr
+	}
 
 	// Copy the event handlers
 	result.EventHandlers = make([]string, 0, len(a.EventHandlers)+len(b.EventHandlers))
@@ -345,6 +424,11 @@ func MergeConfig(a, b *Config) *Config {
 	result.StartJoin = make([]string, 0, len(a.StartJoin)+len(b.StartJoin))
 	result.StartJoin = append(result.StartJoin, a.StartJoin...)
 	result.StartJoin = append(result.StartJoin, b.StartJoin...)
+
+	// Copy the retry join addresses
+	result.RetryJoin = make([]string, 0, len(a.RetryJoin)+len(b.RetryJoin))
+	result.RetryJoin = append(result.RetryJoin, a.RetryJoin...)
+	result.RetryJoin = append(result.RetryJoin, b.RetryJoin...)
 
 	return &result
 }