+++ command/agent/config_test.go
@@ -190,6 +190,17 @@ func TestDecodeConfig(t *testing.T) {
 		t.Fatalf("bad: %#v", config)
 	}
 
+	// tags file
+	input = `{"tags_file": "/some/path"}`
+	config, err = DecodeConfig(bytes.NewReader([]byte(input)))
+	if err != nil {
+		t.Fatalf("err: %s", err)
+	}
+
+	if config.TagsFile != "/some/path" {
+		t.Fatalf("bad: %#v", config)
+	}
+
 	// Discover
 	input = `{"discover": "foobar"}`
 	config, err = DecodeConfig(bytes.NewReader([]byte(input)))
@@ -259,6 +270,84 @@ func TestDecodeConfig(t *testing.T) {
 	if config.TombstoneTimeout != 48*time.Hour {
 		t.Fatalf("bad: %#v", config)
 	}
+
+	// Syslog
+	input = `{"enable_syslog": true, "syslog_facility": "LOCAL4"}`
+	config, err = DecodeConfig(bytes.NewReader([]byte(input)))
+	if err != nil {
+		t.Fatalf("err: %s", err)
+	}
+
+	if !config.EnableSyslog {
+		t.Fatalf("bad: %#v", config)
+	}
+	if config.SyslogFacility != "LOCAL4" {
+		t.Fatalf("bad: %#v", config)
+	}
+
+	// Retry configs
+	input = `{"retry_max_attempts": 5, "retry_interval": "60s"}`
+	config, err = DecodeConfig(bytes.NewReader([]byte(input)))
+	if err != nil {
+		t.Fatalf("err: %s", err)
+	}
+
+	if config.RetryMaxAttempts != 5 {
+		t.Fatalf("bad: %#v", config)
+	}
+
+	if config.RetryInterval != 60*time.Second {
+		t.Fatalf("bad: %#v", config)
+	}
+
+	// Retry configs
+	input = `{"retry_join": ["127.0.0.1", "127.0.0.2"]}`
+	config, err = DecodeConfig(bytes.NewReader([]byte(input)))
+	if err != nil {
+		t.Fatalf("err: %s", err)
+	}
+
+	if len(config.RetryJoin) != 2 {
+		t.Fatalf("bad: %#v", config)
+	}
+
+	if config.RetryJoin[0] != "127.0.0.1" {
+		t.Fatalf("bad: %#v", config)
+	}
+
+	if config.RetryJoin[1] != "127.0.0.2" {
+		t.Fatalf("bad: %#v", config)
+	}
+
+	// Rejoin configs
+	input = `{"rejoin_after_leave": true}`
+	config, err = DecodeConfig(bytes.NewReader([]byte(input)))
+	if err != nil {
+		t.Fatalf("err: %s", err)
+	}
+
+	if !config.RejoinAfterLeave {
+		t.Fatalf("bad: %#v", config)
+	}
+
+	// Rejoin configs
+	input = `{"statsite_addr": "127.0.0.1:8123"}`
+	config, err = DecodeConfig(bytes.NewReader([]byte(input)))
+	if err != nil {
+		t.Fatalf("err: %s", err)
+	}
+
+	if config.StatsiteAddr != "127.0.0.1:8123" {
+		t.Fatalf("bad: %#v", config)
+	}
+}
+
+func TestDecodeConfig_unknownDirective(t *testing.T) {
+	input := `{"unknown_directive": "titi"}`
+	_, err := DecodeConfig(bytes.NewReader([]byte(input)))
+	if err == nil {
+		t.Fatal("should have err")
+	}
 }
 
 func TestMergeConfig(t *testing.T) {
@@ -269,6 +358,7 @@ func TestMergeConfig(t *testing.T) {
 		EventHandlers: []string{"foo"},
 		StartJoin:     []string{"foo"},
 		ReplayOnJoin:  true,
+		RetryJoin:     []string{"zab"},
 	}
 
 	b := &Config{
@@ -286,6 +376,12 @@ func TestMergeConfig(t *testing.T) {
 		RPCAuthKey:            "foobar",
 		DisableNameResolution: true,
 		TombstoneTimeout:      36 * time.Hour,
+		EnableSyslog:          true,
+		RetryJoin:             []string{"zip"},
+		RetryMaxAttempts:      10,
+		RetryInterval:         120 * time.Second,
+		RejoinAfterLeave:      true,
+		StatsiteAddr:          "127.0.0.1:8125",
 	}
 
 	c := MergeConfig(a, b)
@@ -346,12 +442,37 @@ func TestMergeConfig(t *testing.T) {
 		t.Fatalf("bad: %#v", c)
 	}
 
+	if !c.EnableSyslog {
+		t.Fatalf("bad: %#v", c)
+	}
+
+	if c.RetryMaxAttempts != 10 {
+		t.Fatalf("bad: %#v", c)
+	}
+
+	if c.RetryInterval != 120*time.Second {
+		t.Fatalf("bad: %#v", c)
+	}
+
+	if !c.RejoinAfterLeave {
+		t.Fatalf("bad: %#v", c)
+	}
+
+	if c.StatsiteAddr != "127.0.0.1:8125" {
+		t.Fatalf("bad: %#v", c)
+	}
+
 	expected := []string{"foo", "bar"}
 	if !reflect.DeepEqual(c.EventHandlers, expected) {
 		t.Fatalf("bad: %#v", c)
 	}
 
 	if !reflect.DeepEqual(c.StartJoin, expected) {
+		t.Fatalf("bad: %#v", c)
+	}
+
+	expected = []string{"zab", "zip"}
+	if !reflect.DeepEqual(c.RetryJoin, expected) {
 		t.Fatalf("bad: %#v", c)
 	}
 }